chore: move op types; less op cfgs
