+++ Cargo.toml
@@ -248,6 +248,7 @@ syn = "2.0"
 nybbles = "0.2.1"
 smallvec = "1"
 dyn-clone = "1.0.17"
+sha2 = { version = "0.10", default-features = false }
 
 # proc-macros
 proc-macro2 = "1.0"
