+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -23,13 +23,7 @@ use reth_interfaces::{consensus::Consensus, RethResult};
 use reth_node_api::PayloadBuilderAttributes;
 #[cfg(not(feature = "optimism"))]
 use reth_node_ethereum::EthEvmConfig;
-#[cfg(feature = "optimism")]
-use reth_node_optimism::OptimismEvmConfig;
 use reth_payload_builder::database::CachedReads;
-#[cfg(not(feature = "optimism"))]
-use reth_payload_builder::EthPayloadBuilderAttributes;
-#[cfg(feature = "optimism")]
-use reth_payload_builder::OptimismPayloadBuilderAttributes;
 use reth_primitives::{
     constants::eip4844::{LoadKzgSettingsError, MAINNET_KZG_TRUSTED_SETUP},
     fs,
@@ -167,7 +161,7 @@ impl Command {
         let consensus: Arc<dyn Consensus> = Arc::new(BeaconConsensus::new(Arc::clone(&self.chain)));
 
         #[cfg(feature = "optimism")]
-        let evm_config = OptimismEvmConfig::default();
+        let evm_config = reth_node_optimism::OptimismEvmConfig::default();
 
         #[cfg(not(feature = "optimism"))]
         let evm_config = EthEvmConfig::default();
@@ -265,7 +259,7 @@ impl Command {
             Arc::clone(&best_block),
             Bytes::default(),
             #[cfg(feature = "optimism")]
-            OptimismPayloadBuilderAttributes::try_new(
+            reth_node_optimism::OptimismPayloadBuilderAttributes::try_new(
                 best_block.hash(),
                 OptimismPayloadAttributes {
                     payload_attributes: payload_attrs,
@@ -289,9 +283,8 @@ impl Command {
         );
 
         #[cfg(feature = "optimism")]
-        let payload_builder =
-            reth_optimism_payload_builder::OptimismPayloadBuilder::new(self.chain.clone())
-                .compute_pending_block();
+        let payload_builder = reth_node_optimism::OptimismPayloadBuilder::new(self.chain.clone())
+            .compute_pending_block();
 
         #[cfg(not(feature = "optimism"))]
         let payload_builder = reth_ethereum_payload_builder::EthereumPayloadBuilder::default();
