+++ crates/net/network/Cargo.toml
@@ -106,7 +106,6 @@ optimism = [
     "reth-primitives/optimism",
     "reth-provider/optimism",
     "reth-network-api/optimism",
-    "reth-rpc-types/optimism",
 ]
 
 [[bench]]
