+++ crates/node-core/Cargo.toml
@@ -103,7 +103,6 @@ optimism = [
     "reth-provider/optimism",
     "reth-network/optimism",
     "reth-network-api/optimism",
-    "reth-rpc-types/optimism",
     "reth-rpc-types-compat/optimism",
     "reth-auto-seal-consensus/optimism",
     "reth-consensus-common/optimism",
