+++ crates/node-optimism/Cargo.toml
@@ -41,10 +41,8 @@ reth-db.workspace = true
 optimism = [
     "reth-network/optimism",
     "reth-primitives/optimism",
-    "reth-rpc-types/optimism",
     "reth-provider/optimism",
     "reth-rpc-types-compat/optimism",
     "reth-rpc/optimism",
-    "reth-optimism-payload-builder/optimism",
     "reth-revm/optimism",
 ]
