+++ crates/payload/builder/src/lib.rs
@@ -105,7 +105,6 @@ pub mod database;
 pub mod error;
 mod events;
 mod metrics;
-mod optimism;
 mod payload;
 mod service;
 mod traits;
@@ -115,7 +114,6 @@ pub mod noop;
 #[cfg(any(test, feature = "test-utils"))]
 pub mod test_utils;
 
-pub use optimism::{OptimismBuiltPayload, OptimismPayloadBuilderAttributes};
 pub use payload::{EthBuiltPayload, EthPayloadBuilderAttributes};
 pub use reth_rpc_types::engine::PayloadId;
 pub use service::{PayloadBuilderHandle, PayloadBuilderService, PayloadStore};
